% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{addMappackTiles}
\alias{addMappackTiles}
\title{Add local tiles to a leaflet map}
\usage{
addMappackTiles(map, layerId = NULL, group = NULL,
  options = leaflet::tileOptions())
}
\arguments{
\item{map}{a map widget object created from \code{\link{leaflet}()}
'}

\item{layerId}{the layer id}

\item{group}{the name of the grou the newly created layers should belong to}

\item{options}{a list of extra options for tile layers, popups, paths
(circles, rectangles, polygons, ...), or other map elements}
}
\value{
the new \code{map} object
}
\description{
Adds local tiles provided by the \emph{mappack} package to a \code{leaflet} map.
'
}
\references{
The Leaflet API documentation:
  \url{http://leafletjs.com/reference.html}
}
\seealso{
\code{\link{tileOptions}}, \code{\link{WMSTileOptions}},
  \code{\link{popupOptions}}, \code{\link{markerOptions}},
  \code{\link{pathOptions}}
}
