\name{mappack-package}
\alias{mappack-package}
\alias{mappack}
\docType{package}
\title{
\packageTitle{mappack}
}
\description{
\packageDescription{mappack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mappack}
\packageIndices{mappack}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mappack}

Maintainer: \packageMaintainer{mappack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
