% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\docType{data}
\name{tiles}
\alias{tiles}
\alias{mappackTilesURL}
\alias{mappackTilesDir}
\title{Accessors to the tiles}
\format{An object of class \code{character} of length 1.}
\usage{
mappackTilesURL

mappackTilesDir
}
\description{
Use these objects to access the directories for the tiles
in the \code{mappack} package.
}
\keyword{datasets}
